Move data to file
